Refactor backup_management for Zero-Sudo and Odoo 19 compatibility
- Standardized on Zero-Sudo architecture by using module-specific service accounts.
- Updated documentation.html with comprehensive, jargon-free user guides.
- Fixed size_bytes type inconsistency in SQL view to prevent overflow.
- Improved UI tours with structure-agnostic selectors for Odoo 19.
- Verified asynchronous RabbitMQ-offloading pattern and security path validation.
- Cleaned up environment-specific test logic using standard skip decorators.

+++ backup_management/models/sql_views.py
@@ -10,7 +10,7 @@ class BackupLatestSnapshotView(models.Model):
     config_id = fields.Many2one("backup.config", readonly=True)
     snapshot_id = fields.Char(readonly=True)
     start_time = fields.Datetime(readonly=True)
-    size_bytes = fields.Integer(readonly=True)
+    size_bytes = fields.Float(readonly=True)
     status = fields.Char(readonly=True)
 
     def init(self):
